Set up subpage styles
